test/com/blhwong/bsttraversal/ProgramTest.java
package com.blhwong.bsttraversal;

import org.junit.jupiter.api.Test;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

class ProgramTest {

    @Test
    public void TestCase1() {
        var root = new Program.BST(10);
        root.left = new Program.BST(5);
        root.left.left = new Program.BST(2);
        root.left.left.left = new Program.BST(1);
        root.left.right = new Program.BST(5);
        root.right = new Program.BST(15);
        root.right.right = new Program.BST(22);

        List<Integer> inOrder = Arrays.asList(new Integer[] {1, 2, 5, 5, 10, 15, 22});
        List<Integer> preOrder = Arrays.asList(new Integer[] {10, 5, 2, 1, 5, 15, 22});
        List<Integer> postOrder = Arrays.asList(new Integer[] {1, 2, 5, 5, 22, 15, 10});

        assertEquals(Program.inOrderTraverse(root, new ArrayList<Integer>()), inOrder);
        assertEquals(Program.preOrderTraverse(root, new ArrayList<Integer>()), preOrder);
//        assertEquals(Program.postOrderTraverse(root, new ArrayList<Integer>()), postOrder);
    }
}
